+++ .gitignore
@@ -32,6 +32,9 @@ out/
 .vscode/
 
 ### Docker ###
-docker/db/mysql/data/
-docker/db/mysql/init/*
-!docker/db/mysql/init/init.sql
+docker/db/master_db/mysql/data/
+docker/db/master_db/mysql/init/*
+!docker/db/master_db/mysql/init/init.sql
+docker/db/test_db/mysql/data/
+docker/db/test_db/mysql/init/*
+!docker/db/test_db/mysql/init/init.sql
